sync RMF to 9f789d3ea49f7d9e612c9b89e3477164fc3248d3
git-svn-id: https://svn.salilab.org/imp/trunk@15849 e91f8c7b-0e32-4433-a9ac-4502d39a4eab

+++ modules/rmf/dependency/RMF/src/backend/avro/MultipleAvroFileWriter.cpp
@@ -22,7 +22,7 @@ namespace RMF {
     void MultipleAvroFileWriter::set_current_frame(int frame) {
       if (frame == get_current_frame()) return;
       RMF_USAGE_CHECK(frame == ALL_FRAMES
-                      || frame == static_cast<int>(get_frames().size())-2,
+                      || frame == get_frames().size()-2,
                       "Bad frame set");
       MultipleAvroFileBase::set_current_frame(frame);
     }
